api/requirements.txt
# API Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Queue
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
pycountry==23.12.11

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
mypy==1.7.1